requirements.txt
asgiref==3.5.0
Django==4.0.4
django-debug-toolbar==3.2.4
django-filter==21.1
mysqlclient==2.1.0
Pillow==9.1.0
sqlparse==0.4.2
